Refactor to simplify instructions, change datatype to Int64.
Refactor instructions to provide all parameter values from within the base class rather than complex intermediate classes.

Change int data types to Int64 from Int32.

+++ Advent of Code/Day 05/Day_05.cs
@@ -16,14 +16,14 @@ namespace Advent_of_Code.Day_05
             List<int> program = SharedLibrary.FileParser.GetIntCodeFromFile(@"..\..\..\Day 05\input.txt");
 
             Console.WriteLine($"=====Day 5, part 1=====");
-            IntcodeVirtualMachine intcodeVirtualMachine = new IntcodeVirtualMachine(program, new StaticInputProvider(new List<int>() { 1 }), new FileOutputProvider("Day5Part1.txt"));
+            IntcodeVirtualMachine intcodeVirtualMachine = new IntcodeVirtualMachine(program, new StaticInputProvider(new List<long>() { 1 }), new FileOutputProvider("Day5Part1.txt"));
             intcodeVirtualMachine.Run();
             Console.WriteLine($"Output = {System.IO.File.ReadAllText("Day5Part1.txt")}");
 
 
             program = SharedLibrary.FileParser.GetIntCodeFromFile(@"..\..\..\Day 05\input.txt");
             Console.WriteLine($"=====Day 5, part 2=====");
-            intcodeVirtualMachine = new IntcodeVirtualMachine(program, new StaticInputProvider(new List<int>() { 5 }), new FileOutputProvider("Day5Part2.txt"));
+            intcodeVirtualMachine = new IntcodeVirtualMachine(program, new StaticInputProvider(new List<long>() { 5 }), new FileOutputProvider("Day5Part2.txt"));
             intcodeVirtualMachine.Run();
             Console.WriteLine($"Output = {System.IO.File.ReadAllText("Day5Part2.txt")}");
