+++ Advent of Code/Day 07/Amplifier.cs
@@ -9,20 +9,20 @@ namespace Advent_of_Code.Day_07
 {
     public class Amplifier
     {
-        private readonly int _input;
-        private readonly int _phase;
+        private readonly long _input;
+        private readonly long _phase;
         private readonly IntcodeVirtualMachine _intcodeVirtualMachine;
-        private readonly List<int> _program;
+        private readonly List<long> _program;
         private readonly string _filename;
 
-        public Amplifier(int input, int phase, List<int> program, IInputProvider inputProvider = null, IOutputProvider outputProvider = null)
+        public Amplifier(long input, long phase, List<long> program, IInputProvider inputProvider = null, IOutputProvider outputProvider = null)
         {
             this._input = input;
             this._phase = phase;
             _program = program;
             _filename = Path.GetTempFileName();
 
-            inputProvider ??= new StaticInputProvider(new List<int>() { _phase, _input });
+            inputProvider ??= new StaticInputProvider(new List<long>() { _phase, _input });
             outputProvider ??= new FileOutputProvider(_filename);
 
             _intcodeVirtualMachine = new IntcodeVirtualMachine(_program, inputProvider, outputProvider);
